+++ app/models.py
@@ -4,7 +4,6 @@ import enum
 import logging
 import uuid
 from urllib.parse import parse_qs
-
 import github
 import lib
 import requests
@@ -12,107 +11,12 @@ from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.db import models
 from django.utils import timezone
+from allauth.socialaccount import models as social_models
 
 logger = logging.getLogger(__name__)
 
 
-class Token(models.Model):
-    """
-    This model takes care of the tokens generated by the Github applications.
-    """
-
-    token = models.CharField(max_length=255)
-    created_at = models.DateTimeField(auto_now_add=True)
-    expires_at = models.DateTimeField()
-
-    class Meta:
-        abstract = True
-
-
-class GithubUser(models.Model):
-    user = models.OneToOneField(
-        get_user_model(), on_delete=models.CASCADE, related_name="github_user"
-    )
-    account_name = models.CharField(max_length=200)
-    account_id = models.BigIntegerField()
-    account_type = models.CharField(max_length=20)
-    avatar_url = models.CharField(max_length=200)
-    created_at = models.DateTimeField(auto_now_add=True)
-
-    @property
-    def access_token(self):
-        try:
-            return (
-                self.access_tokens.exclude(expires_at__lt=timezone.now())
-                .latest("expires_at")
-                .token
-            )
-        except:
-            return None
-
-    def get_active_access_token(self):
-        access_token = self.access_token
-        if access_token is None:
-            return self.get_new_tokens()[0]
-        return access_token
-
-    @property
-    def refresh_token(self):
-        try:
-            return (
-                self.refresh_tokens.exclude(expires_at__lt=timezone.now())
-                .latest("expires_at")
-                .token
-            )
-        except:
-            return None
-
-    def process_token_response(self, token_response):
-        now = timezone.now()
-        logger.info(token_response)
-        data = parse_qs(token_response)
-        access_token = GithubUserAccessToken.objects.create(
-            token=data["access_token"][0],
-            github_user=self,
-            expires_at=now + datetime.timedelta(seconds=int(data["expires_in"][0])),
-        )
-        refresh_token = GithubUserRefreshToken.objects.create(
-            token=data["refresh_token"][0],
-            github_user=self,
-            expires_at=now
-            + datetime.timedelta(seconds=int(data["refresh_token_expires_in"][0])),
-        )
-        return (access_token.token, refresh_token.token)
-
-    def get_new_tokens(self):
-        refresh_token = self.refresh_tokens.exclude(
-            expires_at__lt=timezone.now()
-        ).latest("expires_at")
-        response = requests.post(
-            "https://github.com/login/oauth/access_token",
-            data={
-                "refresh_token": refresh_token.token,
-                "grant_type": "refresh_token",
-                "client_id": settings.GITHUB_CREDS["client_id"],
-                "client_secret": settings.GITHUB_CREDS["client_secret"],
-            },
-        )
-        return self.process_token_response(response.content.decode())
-
-
-class GithubUserAccessToken(Token):
-    github_user = models.ForeignKey(
-        GithubUser, on_delete=models.CASCADE, related_name="access_tokens"
-    )
-
-
-class GithubUserRefreshToken(Token):
-    github_user = models.ForeignKey(
-        GithubUser, on_delete=models.CASCADE, related_name="refresh_tokens"
-    )
-
-
-class GithubAppInstallation(models.Model):
+class AppInstallation(models.Model):
     """
     This model takes care of the Github installations.
     """
@@ -122,8 +26,9 @@ class GithubAppInstallation(models.Model):
         UNINSTALLED = "UNINSTALLED", "Uninstalled"
         SUSPENDED = "SUSPENDED", "Suspended"
 
+    social_app = models.ForeignKey(social_models.SocialApp, on_delete=models.PROTECT)
     installation_id = models.BigIntegerField(unique=True, db_index=True)
-    creator = models.ForeignKey(GithubUser, on_delete=models.PROTECT)
+    creator = models.ForeignKey(get_user_model(), on_delete=models.PROTECT)
     state = models.CharField(max_length=20, choices=InstallationState.choices)
     account_name = models.CharField(max_length=200)
     account_id = models.BigIntegerField()
@@ -131,71 +36,43 @@ class GithubAppInstallation(models.Model):
     avatar_url = models.CharField(max_length=200)
     created_at = models.DateTimeField(auto_now_add=True)
 
+    class Meta:
+        unique_together = ("social_app", "installation_id")
+
     def __str__(self) -> str:
-        return f"{self.account_name}[{self.installation_id}] Owner: {self.creator.account_name}"
-
-    @property
-    def access_token(self):
-        try:
-            return (
-                self.tokens.exclude(expires_at__lte=timezone.now())
-                .latest("created_at")
-                .token
-            )
-        except GithubInstallationToken.DoesNotExist:
-            self.update_token()
-            return self.get_latest_active_token()
-
-    def update_token(self):
-        """
-        Fetches the latest access token for the installation. Returns true/false depending on whether
-        the requested token has been updated
-        """
-        # User token not required for this step
-        github_installation_manager = lib.GithubInstallationManager(
-            installation_id=self.installation_id, user_token=""
-        )
-        token, expires_at = github_installation_manager.get_installation_access_token()
-        return GithubInstallationToken.objects.get_or_create(
-            token=token,
-            expires_at=expires_at,
-            github_app_installation=self,
-        )[1]
+        return f"{self.account_name}[{self.installation_id}] Owner: {self.creator.username}"
 
 
-class GithubAppUser(models.Model):
-    installation = models.ForeignKey(GithubAppInstallation, on_delete=models.CASCADE)
-    github_user = models.ForeignKey(GithubUser, on_delete=models.CASCADE)
+class AppUser(models.Model):
+    installation = models.ForeignKey(AppInstallation, on_delete=models.CASCADE)
+    user = models.ForeignKey(get_user_model(), on_delete=models.CASCADE)
     created_at = models.DateTimeField(auto_now_add=True)
 
     class Meta:
-        unique_together = ("installation", "github_user")
-
-    def __str__(self) -> str:
-        return f"{self.github_user.account_name}[{self.installation.installation_id}]"
+        unique_together = ("installation", "user")
 
 
-class GithubInstallationToken(Token):
-    github_app_installation = models.ForeignKey(
-        GithubAppInstallation, on_delete=models.CASCADE, related_name="tokens"
-    )
-
-
-class GithubRepository(models.Model):
+class Repository(models.Model):
     repo_id = models.BigIntegerField()
     repo_name = models.CharField(max_length=150)
     repo_full_name = models.CharField(max_length=200)
-    owner = models.ForeignKey(GithubAppInstallation, on_delete=models.PROTECT)
+    app = models.ForeignKey(social_models.SocialApp, on_delete=models.PROTECT)
     created_at = models.DateTimeField(auto_now_add=True)
 
     class Meta:
-        unique_together = ("repo_id", "owner")
+        unique_together = ("repo_id", "app")
 
     def __str__(self) -> str:
         return self.repo_full_name
 
-    def get_url(self):
-        return f"https://github.com/{self.repo_full_name}/pulls/"
+
+class UserRepoAccess(models.Model):
+    repository = models.ForeignKey(Repository, on_delete=models.CASCADE)
+    user = models.ForeignKey(get_user_model(), on_delete=models.CASCADE)
+    created_at = models.DateTimeField(auto_now_add=True)
+
+    class Meta:
+        unique_together = ("repository", "user")
 
 
 class GithubRepoMap(models.Model):
@@ -208,12 +85,12 @@ class GithubRepoMap(models.Model):
         FULL = "FULL", "Full"
         PARTIAL = "PARTIAL", "Partial"
 
-    integration = models.ForeignKey(GithubAppInstallation, on_delete=models.PROTECT)
+    integration = models.ForeignKey(AppInstallation, on_delete=models.PROTECT)
     code_repo = models.ForeignKey(
-        GithubRepository, on_delete=models.PROTECT, related_name="code_repos"
+        Repository, on_delete=models.PROTECT, related_name="code_repos"
     )
     documentation_repo = models.ForeignKey(
-        GithubRepository, on_delete=models.PROTECT, related_name="documentation_repos"
+        Repository, on_delete=models.PROTECT, related_name="documentation_repos"
     )
     integration_type = models.CharField(max_length=20, choices=IntegrationType.choices)
     created_at = models.DateTimeField(auto_now_add=True)
@@ -222,7 +99,7 @@ class GithubRepoMap(models.Model):
         unique_together = ("integration", "code_repo", "documentation_repo")
 
 
-class GithubPullRequest(models.Model):
+class PullRequest(models.Model):
     pr_id = models.BigIntegerField()
     pr_number = models.BigIntegerField()
     pr_head_commit_sha = models.CharField(max_length=40)
@@ -238,7 +115,7 @@ class GithubPullRequest(models.Model):
     pr_merged = models.BooleanField(default=False)
     pr_owner_username = models.CharField(max_length=100)
     updated_on = models.DateTimeField(auto_now=True)
-    repository = models.ForeignKey(GithubRepository, on_delete=models.PROTECT)
+    repository = models.ForeignKey(Repository, on_delete=models.PROTECT)
 
     def __str__(self) -> str:
         return f"{self.repository.repo_full_name}/{self.pr_number} @ {self.pr_head_commit_sha[:7]}"
@@ -267,12 +144,12 @@ class MonitoredPullRequest(models.Model):
         MANUAL_APPROVAL = "MANUALLY_APPROVED", "Manual Approval"
 
     code_pull_request = models.OneToOneField(
-        GithubPullRequest,
+        PullRequest,
         on_delete=models.CASCADE,
         related_name="monitored_code",
     )
     documentation_pull_request = models.ForeignKey(
-        GithubPullRequest,
+        PullRequest,
         on_delete=models.CASCADE,
         blank=True,
         null=True,
@@ -283,7 +160,7 @@ class MonitoredPullRequest(models.Model):
         choices=PullRequestStatus.choices,
         default=PullRequestStatus.NOT_CONNECTED,
     )
-    integration = models.ForeignKey(GithubAppInstallation, on_delete=models.CASCADE)
+    integration = models.ForeignKey(AppInstallation, on_delete=models.CASCADE)
 
     class Meta:
         unique_together = ("code_pull_request", "documentation_pull_request")
@@ -340,7 +217,7 @@ class PrApproval(models.Model):
         super().save(*args, **kwargs)
 
 
-class GithubCheckRun(models.Model):
+class PullRequestCheck(models.Model):
     run_id = models.BigIntegerField(blank=True, null=True)
     unique_id = models.UUIDField(default=uuid.uuid4)
     # Head of the request when it was run
@@ -385,20 +262,6 @@ class GithubCheckRun(models.Model):
             )
             self.run_id = check_run_instance.id
         super().save(*args, **kwargs)
-
-
-class GithubLoginState(models.Model):
-    """
-    This model is used to store the state of the login process. Along with it,
-    it also stores the redirect URL for the state.
-    """
-
-    state = models.UUIDField(default=uuid.uuid4)
-    redirect_url = models.URLField()
-    created_at = models.DateTimeField(auto_now_add=True)
-
-    def __str__(self) -> str:
-        return self.state.__str__()
 
 
 class GithubMarketplaceEvent(models.Model):
